+++ dace/libraries/linalg/nodes/tensordot.py
@@ -4,7 +4,7 @@ import dace
 import dace.libraries.linalg.environments as environments
 
 from dace import library, nodes, properties
-from dace.data import _prod
+from dace.utils import prod as _prod
 from dace.libraries.blas import blas_helpers
 from dace.symbolic import symstr
 from dace.transformation.transformation import ExpandTransformation
