+++ dace/optimization/data_layout_tuner.py
@@ -5,7 +5,7 @@ import enum
 import copy
 import itertools
 
-from typing import Generator, Optional, Tuple, Dict, List, Sequence, Set
+from typing import Any, Generator, Optional, Tuple, Dict, List, Sequence, Set
 
 from dace import data as dt, SDFG, dtypes
 from dace.optimization import cutout_tuner
@@ -18,6 +18,11 @@ try:
     from tqdm import tqdm
 except (ImportError, ModuleNotFoundError):
     tqdm = lambda x, **kwargs: x
+
+try:
+    from numpy.typing import ArrayLike
+except ImportError:
+    ArrayLike = Any  # type: ignore
 
 
 class TuningGroups(enum.Enum):
@@ -111,7 +116,7 @@ class DataLayoutTuner(cutout_tuner.CutoutTuner):
         cutout.instrument = self.instrument
 
         # Prepare original arguments to sub-SDFG from instrumented data report
-        arguments: Dict[str, dt.ArrayLike] = {}
+        arguments: Dict[str, ArrayLike] = {}
         for cstate in cutout.nodes():
             for dnode in cstate.data_nodes():
                 if cutout.arrays[dnode.data].transient:
